MAKE-581: InfoHasMatchingID occasionally fails for Jasper process tests (#25)
This issue was being caused by an inherent race, that is the fault of
some bad test code. No user should be expecting consistent behavior from
proc.Info() if they have not called proc.Wait() to ensure that the
process has finished executing. If they do not, it could potentially
give back an incomplete process info since Wait() had not yet been
called. The test now does this, so this failure should no longer arise.

This commit also "fixes" a single LoC where we were trying to wrap an
error even when there may not have been an error. This has been wrapped
in an `if err != nil` check.